requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
exceptiongroup==1.3.0
greenlet==3.2.4
iniconfig==2.1.0
loguru==0.7.2
numpy==2.2.6
packaging==25.0
pandas==2.2.2
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.10.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.32
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
